Added the implementation of the code blocks
Created multiple classes:
CodeBlock - A object that is able to be dragged and dropped
SpawnCodeBlock - A object that when pressed could create a new Code Block

Created a test code challenge scene where a player can spawn new test code blocks and drag it to an starting point. Once dragged to the starting point, it will be added to a list and position will be changed to align with others. If a code block is dragged out, it will be removed from the list and all blocks positions will update.

To enter the Test Code Challenge, after running the game go right to a Godot icon, when your player is over the icon hit "space"