Added health bar
Added a HUD scene that features a health bar, and two buttons to add or remove health. If you set the health mode to empty or simple, ensure that max_health matches the number of hearts. If you set it to partial, ensure that the max_health is double the amount of hearts.

Function update_health() can be used to change the player's health, from damage sources or healing sources