Adding to User_Documentation.md (Description below)
Added more descriptive measures to the "running the software" process with VS Code integration, listing every step one can take to be able to run the project using VS Code and GitHub Desktop. Added more to the "how to use the software" section, adding descriptions to various components like "load game" and what happens when you click those elements from the main menu options. Added to the "coding challenges" section to reiterate that players will only have 3 lives before having to restart the process of interacting with npcs to try answering again. Added more to the "missing" section that says we may try to implement a leadership board or component that allows users to see their metrics (challenges they got correct or wrong, etc, attempts, time etc) also added to the "how to report a bug section" which specified that the users should be descriptive about the steps they took in the game that lead them to discover a bug.